python3-aiohttp: add missing RDEPENDS on python3-typing-extensions
Since aiohttp #5078 PR was backported to 3.7 release
(https://github.com/aio-libs/aiohttp/pull/5079), aiohttp depends on
'typing_extensions' python module [1], provided by python3-typing-extensions
package in meta-python.

[1] https://github.com/aio-libs/aiohttp/blob/v3.7.2/setup.py#L74

This bug must be fixed in meta-python.
This patch can be removed as soon as python3-aiohttp is fixed upstream.

As we do not hard-depend on meta-python, we put the .bbappend file in
dynamic-layers.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>